+++ src/app/core/services/index.ts
@@ -8,4 +8,3 @@ export * from './media.service';
 export * from './event.service';
 export * from './clipboard.service';
 export * from './firebase.service';
-export * from './tmp.service';
