+++ src/app/routes/create/create.component.html
@@ -1,6 +1,6 @@
 <div class="page">
   <div class="id">
-    <input type="text" readonly="true" [value]="dataService.id">
+    <input type="text" readonly="true" [value]="id">
     <button (click)="randomize()">Randomize</button>
   </div>
   <div>
